feat: restructure files
