+++ mfins-doctor/go.mod
@@ -1,0 +1,3 @@
+module github.com/metakgp/mfins/mfins-doctor
+
+go 1.23.1
